Make compiler fail cleanly if malloc fails
We now report the failure and exit.  Previously the NULL return from
malloc wasn't checked for so we'd typically segfault.

+++ compiler/header.h
@@ -146,7 +146,7 @@ extern void disable_token(struct tokeniser * t, int code);
 extern void close_tokeniser(struct tokeniser * t);
 
 extern int space_count;
-extern void * check_malloc(int n);
+extern void * check_malloc(size_t n);
 extern void check_free(void * p);
 
 struct node;
